Move compiler/parser related functionality from AST to base classes
Part of adopting new frame design for bytecode interpreter.
